+++ mne/io/base.py
@@ -11,6 +11,7 @@ import copy
 from copy import deepcopy
 import os
 import os.path as op
+import warnings
 
 import numpy as np
 
@@ -365,7 +366,7 @@ class BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin,
         self._projectors = list()
         self._projector = None
         self._dtype_ = dtype
-        self.annotations = None
+        self.set_annotations(None)
         # If we have True or a string, actually do the preloading
         self._update_times()
         if load_from_disk:
@@ -666,6 +667,13 @@ class BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin,
 
     @annotations.setter
     def annotations(self, annotations, emit_warning=True):
+        warnings.warn('setting the annotations attribute by assignation is'
+                      ' deprecated since 0.17, and would be removed in 0.19.'
+                      ' Please use raw.set_annotations() instead.',
+                      category=DeprecationWarning)
+        self.set_annotations(annotations, emit_warning=emit_warning)
+
+    def set_annotations(self, annotations, emit_warning=True):
         """Setter for annotations.
 
         This setter checks if they are inside the data range.
@@ -677,19 +685,24 @@ class BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin,
         emit_warning : bool
             Whether to emit warnings when limiting or omitting annotations.
         """
+        new_annot = None
         if annotations is not None:
             if not isinstance(annotations, Annotations):
                 raise ValueError('Annotations must be an instance of '
                                  'mne.Annotations. Got %s.' % annotations)
+
+            new_annot = annotations.copy()
             meas_date = _handle_meas_date(self.info['meas_date'])
-            if annotations.orig_time is not None:
-                offset = (annotations.orig_time - meas_date -
+
+            if new_annot.orig_time is not None:
+                offset = (new_annot.orig_time - meas_date -
                           self.first_samp / self.info['sfreq'])
             else:
                 offset = 0
+
             omit_ind = list()
             omitted = limited = 0
-            for ind, onset in enumerate(annotations.onset):
+            for ind, onset in enumerate(new_annot.onset):
                 onset += offset
                 if onset > self.times[-1]:
                     omitted += 1
@@ -697,29 +710,28 @@ class BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin,
                     logger.debug('Omitting %d @ %s > %s'
                                  % (ind, onset, self.times[-1]))
                 elif onset < self.times[0]:
-                    if onset + annotations.duration[ind] < self.times[0]:
+                    if onset + new_annot.duration[ind] < self.times[0]:
                         omitted += 1
                         omit_ind.append(ind)
                         logger.debug('Omitting %d @ %s < %s'
                                      % (ind, onset, self.times[0]))
                     else:
                         limited += 1
-                        duration = annotations.duration[ind] + onset
-                        annotations.duration[ind] = duration
-                        annotations.onset[ind] = self.times[0] - offset
-                elif (onset + annotations.duration[ind] >
+                        duration = new_annot.duration[ind] + onset
+                        new_annot.duration[ind] = duration
+                        new_annot.onset[ind] = self.times[0] - offset
+                elif (onset + new_annot.duration[ind] >
                       self.times[-1] + 1.1 / self.info['sfreq']):
                     # We have to permit onset+duration to appear to go one past
                     # the last sample in order to actually include the last
                     # sample...
                     limited += 1
-                    annotations.duration[ind] = (self.times[-1] +
-                                                 1. / self.info['sfreq'] -
-                                                 onset)
-            annotations.onset = np.delete(annotations.onset, omit_ind)
-            annotations.duration = np.delete(annotations.duration, omit_ind)
-            annotations.description = np.delete(annotations.description,
-                                                omit_ind)
+                    new_annot.duration[ind] = (self.times[-1] +
+                                               1. / self.info['sfreq'] -
+                                               onset)
+            new_annot.onset = np.delete(new_annot.onset, omit_ind)
+            new_annot.duration = np.delete(new_annot.duration, omit_ind)
+            new_annot.description = np.delete(new_annot.description, omit_ind)
             if emit_warning:
                 if omitted > 0:
                     warn('Omitted %s annotation(s) that were outside data '
@@ -727,7 +739,8 @@ class BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin,
                 if limited > 0:
                     warn('Limited %s annotation(s) that were expanding '
                          'outside the data range.' % limited)
-        self._annotations = annotations
+
+        self._annotations = new_annot
 
     def __del__(self):  # noqa: D105
         # remove file for memmap
@@ -1611,11 +1624,10 @@ class BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin,
         self._update_times()
 
         if self.annotations is not None:
-            annotations = self.annotations
-            # XXX there might be a cleaner way to do this someday
             if self.annotations.orig_time is None:
                 self.annotations.onset -= tmin
-            BaseRaw.annotations.fset(self, annotations, emit_warning=False)
+            # now call setter to filter out annotations outside of interval
+            self.set_annotations(self.annotations, False)
 
         return self
 
@@ -2036,7 +2048,7 @@ class BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin,
         self._update_times()
         if annotations is None:
             annotations = Annotations([], [], [])
-        self.annotations = annotations
+        self.set_annotations(annotations)
         for edge_samp in edge_samps:
             onset = _sync_onset(self, (edge_samp) / self.info['sfreq'], True)
             self.annotations.append(onset, 0., 'BAD boundary')
@@ -2163,6 +2175,13 @@ class _RawShell():
     @property
     def n_times(self):  # noqa: D102
         return self.last_samp - self.first_samp + 1
+
+    @property
+    def annotations(self):  # noqa: D102
+        return self._annotations
+
+    def set_annotations(self, annotations):
+        self._annotations = annotations
 
 
 ###############################################################################