+++ mne/preprocessing/tests/test_ica.py
@@ -91,7 +91,7 @@ def test_ica_full_data_recovery(method):
     data = raw._data[:n_channels].copy()
     data_epochs = epochs.get_data()
     data_evoked = evoked.data
-    raw.annotations = Annotations([0.5], [0.5], ['BAD'])
+    raw.set_annotations(Annotations([0.5], [0.5], ['BAD']))
     methods = [method]
     for method in methods:
         stuff = [(2, n_channels, True), (2, n_channels // 2, False)]
@@ -334,7 +334,7 @@ def test_ica_additional(method):
     stop2 = 500
     raw = read_raw_fif(raw_fname).crop(1.5, stop).load_data()
     raw.del_proj()  # avoid warnings
-    raw.annotations = Annotations([0.5], [0.5], ['BAD'])
+    raw.set_annotations(Annotations([0.5], [0.5], ['BAD']))
     # XXX This breaks the tests :(
     # raw.info['bads'] = [raw.ch_names[1]]
     test_cov = read_cov(test_cov_name)